grpc/testserver: close always stops server (#83)
Previously, closing the test server checked for errors when closing the
test connection before stopping the server.

However, closing the connection will return errors for cases where we
still need to shut down the server, for example, if the client has
closed its connection already.

* In order to keep our long-lived connections alive, we need a keepalive mechanism.
* There are really 3 layers keepalives can be injected: TCP, HTTP2, and gRPC application layer.
* TCP keepalives are enabled by default through gRPC, but out connections go through Amazon ELBs, which ignores TCP keepalives for maintaining long-lived connections.

HTTP2 `PING` frames are another alternative, but there is no clear or documented way to access this layer of the connections through gRPC.

That leaves the application layer.

gRPC ships with a [health checking service](https://github.com/grpc/grpc/blob/master/doc/health-checking.md) which seems appropriate to our needs.

Adding health checks requires some changes in both the domain streamer and funnel services.

The streaming clients perform periodic healthchecks on the connection. The healthchecks will be sent if nothing is streamed for 30s, which should fit comfortably under the ELB's 60s timeout.

If for some reason the server responds with `NOT_SERVING` or `UNKNOWN` status, the client shuts down.

* cancel stream context

When `StreamingClient.Close` is called, we previously only closed the
gRPC connection and allowed internal gRPC mechanisms to shut down the
stream.

However this does not seem to be a reliable method of terminating a
stream. At least in tests with graceful shutdown, there seem to be race
conditions which cause the stream not to terminate.

Now `StreamingClient.Close` also cancels the stream's context, which
immediately shuts down the stream.

This allows us to turn back on the test server's `Dialer` method, which
was disabled due to flakiness in the tests around shutdown.

It also adds a timeout to graceful stop, so that tests will not hang
when it times out.

* Fix missing cancel in error case