lib: Add support for l2met-formatted logs with go-kit's metrics package (#154)
## Reasons for making these changes

I'd like to experiment with go-kit's metrics package to see how it could work for us. Ideally, we'll report our metrics directly to Librato, but in the interest of both making a small change and in learning how reporting works, I thought it'd be interesting to explore adding support for l2met.

## Description of changes

A new `lib/kit/metrics/l2met` package provides an `L2met` struct, based on the graphite implementation for go-kit itself, to generate and emit count, gauge and histogram metrics. It vendors `github.com/go-kit/kit/metrics` and related packages.

## Metadata

- [Pivotal Tracker](https://www.pivotaltracker.com/story/show/139246711)