grpcmetrics: use the registry (#467)
* grpcmetrics: use the registry

== Reasons for doing this

So we can potentially switch to different implementations of a
gokit/metrics.Provider. Currently event gateway is experimenting with
running the librato Provider.

== Description of changes

Swap out uses of the Provider directly with a prefixedRegistry.

* Remove prefixProvider

* Rename reg to r0 / r1

* testmetrics: NewXYZ should only be callable once

* Remove vestigial name

* Move registry to a private, grpcmetrics only thing

As this was being discussed, the value of a Registry was asked
and in retrospect, making it public is premature, since today
grpcmetrics is the only thing that requires this.