grpc/grpchttp: add package for bridging grpc and http (#721)
* grpc/grpchttp: add package for bridging grpc and http

We'll need to manage some transformations in kpi-http-gateway between
HTTP and gRPC. `grpc-gateway` provides most of the front-end
machinery.

But we'll also need to translate back-end errors -- for example, a 404
from psmgr -- into correct error codes for the front-end. We'll also
need some helpers to do things like inject Request-Id on incoming HTTP
requests into the grpc context. The `grpchttp` package will be
responsible for these tasks.

This adds `grpchttp.CodeFromHTTPStatus` which is the reverse of
[runtime.HTTPStatusFromCode](http://godoc.org/github.com/grpc-ecosystem/grpc-gateway/runtime#HTTPStatusFromCode).

* Fixup

* Use http.StatusTooManyRequests