grpc: add requestid package and grpchttp annotation (#722)
* grpc: add requestid package and grpchttp annotation

Incoming requests to kpi-http-gateway include a Request-Id in the
headers which should be propagated to backing services.

This adds a new `grpc/requestid` package and extends `grpchttp` to
handle incoming request IDs.

grpchttp.RequestIDAnnotator is a small wrapper for use with
grpc-gateway.

The packages are separated out, because the request ID will propagate
automatically to gRPC backing services through the metadata. That means
the request ID set by RequestIDAnnotator will be available in
release-manager or domain-streamer via `requestid.FromContext`.

We can propagate with something like:

* Use balancer package