Update to logrus 1.0 with new import name (#738)
logrus has been renamed from Sirupsen/logrus to sirupsen/logrus. We need
to use the new name to be able to continue to use it going forward, as
well as to bring in other community packages which have already been
updated.

This updates logrus and rollrus, and changes everywhere that was
Sirupsen/logrus to sirupsen/logrus.