grpc/grpcserver: Add support for nested logging tags (#891)
The `lib/grpc/grpcserver` package logs debugging data for types that satisfy the loggable interface, ie, they have a `LoggingTags() map[string]interface{}` method. The `protoc-gen-loggingtags` program added in #800 adds support to generate `LoggingTags() map[string]interface{}` methods for messages defined in protobufs, which is intended to make debug logging essentially free, but missing support to handle nested messages in the log generation path makes the overall functionality less useful than it could be.

The logic to extract data from types that satisfy the loggable interface has been extended to handle nested types. That is, if a value in the `map[string]interface{}` returned by `LoggingTags()` implements the loggable interface, it will be logged with a key prefix made from the current scope and key.

The comment at the top of generated code has also been modified to match the conventions used by other protobuf generators. This causes the diff for generated code to be hidden in GitHub's *Files changes* tab.