Make gRPC servers configurable. (#900)
The "standard" options we use for gRPC are fixed, and passed in
as an argument list. This makes any kind of optional configuration
hard, and makes adding/removing customizable options configure a
refactor.

Set up an options mechanism similar to what gRPC itself uses, and
use this to configure the logging and metrics items. Adjust current
consumers to use this.