gRPC Token Authorizer for Heroku API (#902)
For user-facing services, a convenient auth source is the Heroku
API. This adds a tokenauth.Authorizer that can proxy credentials
extracted from the users ~/.netrc to the Heroku API to verify
their validity, and check the domain the user belongs to for
authorization purposes. The user's email and API UUID are then
embedded in the context for further use.