dyno-funnel/api: wire up /dynos endpoint (#1176)
* dyno-funnel/api: wire up /dynos endpoint

== Rationale

We want the dyno funnel API server to be able to serve a stream of dynos
given all the configured psmgr common spaces.

The model for serving is also to make the stream resilient to individual
failures (e.g. if va01 is faulty for a given time interval (125s max is
what we live with) then the rest of the other streams should continue to
function. (This just uses the previously built MultiDynoClient)

== Description

- Implement concept of Dialer in api.
  The Dialer just returns a client given ctx / since. This makes testing
  easier.
- Wire up api.New with a Dialer
- Add all necessary configuration parameters in
  cmd/cedar/internal/dynofunnelapicmd/config.go
  We also move all the config related responsibilites here, e.g.
  InstanceName, and PsmgrSourceURLs

* Prevent race issues, refactor a bit

* Fix since test in multi client