lib/grpc/grpcserver: refactor in-process gRPC servers (#1218)
This continues on the work from #1199 to rework how we handle in-process
gRPC servers, for use in tests and with grpc-gateway.

This unifies NewInProcess, NewStandardInProcess, and testserver.New into
a single implementation. All uses of gRPC servers in cedar should now be
ready for upgrade to the new version of grpc-go (which can be verified
in a followup PR).

"InProcess" is now Local, as it reads a bit better and has a natural
name on the left-hand side.