grpcmetrics: cleanup cancelation handling (#1270)
grpcmetrics was previously inconsistent in how it handled cancelation
errors versus others. In particular it reported context cancelation as
an "unknown" response and then measured on a `context-canceled-errors`
metric.

This updates cancelation errors to be counted as "canceled" responses
without special handling in the error case. It also now handles gRPC
cancel errors in addition to context cancel errors.