A bunch of cleanups (see comment)
Document everything more, including usage of all exported bits.

Reduce the use of global variables down to the started flag and it's mutex.

Report is now synchronous, pushing that off to the caller if they directly use
Report. onload is unaffected. Cleaned up examples to reflect this.

Fixed the package name declaration in onload (was onloads by accident).

Renamed a few functions and variables for clarity.