misc: prefer stdlib context to net/context (#1438)
We previously had to use net/context for methods which implemented grpc
interfaces, and stdlib's context for everything else. This required some
massaging of otherwise-identical contexts to make everything compatible.

Since we are now using Go 1.9, however, we no longer need to use
net/context for grpc! That's because net/context's Context type is a
[type alias for context.Context under 1.9][alias].

This removes all uses of net/context from heroku/cedar outside of vendor
and grpc-generated files.

It also re-goimports a few files which had inconsistent groupings of
imports.

  [alias]: https://github.com/golang/net/blob/master/context/go19.go#L15