misc: mark shared test helpers as helpers (#1443)
We're running Go 1.9, which adds a `Helper()` method to testing.T and
other testing types to mark functions as helpers. These functions will
be omitted from the error trace if they fail. For example, in
calls to `testMetricsProvider.CheckCounter()` would previously fail
with something like:

    provider.go:105: snapshotter.total-domains = 1, want 2

Where `provider.go:105` is the location of the `t.Fatal` call in
`CheckCounter` itself. These checks will now instead point to the line
which calls `CheckCounter`:

    snapshot_test.go:138: snapshotter.total-domains = 1, want 2

The kafka testconsumer is also updated to remove its dependency on
testing.T since it was unused.