mustcert: Slightly more flexible cert generation
Rename "hostname" to "commonName", to make what it is a little clearer for
non-serving usages. Add a optional function arg to the newCert and Leaf
methods, that takes a copy of the Cert that it can modify. This can be used to
add additional fields while keeping a simple method signature.