Add UniqueCounter to the Provider
Implement, and test UniqueCounter in the context of the Librato
Provider. Also, provide a discarded counter for when that's
required.

This actually makes the Provider a bit awkward to use, as it requires
referencing the UniqueCounter type, too, but, go-kit doesn't provide
a suitable type with the ability to add the value to count...