Remove token+heroku auth, set interceptor direct (#2690)
To support kubespace/kubemgr, a token authentication method and
implementation against the Heroku API was added.

In the most recent @heroku/runtime-nsotb work that has been removed from
those services, in favour of using "runtime id"
(https://github.com/heroku/runtime-id). This passes the tokens
differently, more inline with the standard oauth2 method.

As this was the only service using that authorization method, remove
from this codebase. To support the runtime-id usage, switch the old
token auth server option to be an option that just takes a set of
interceptors - this will allow auth to be added in a generic fashion.
Even though technically any kind of interceptor could be put here I
chose to call the option "auth", to avoid dealing with configurable
ordering.

The auth interceptors for runtime id still reside externally. With the
current use of the service and need to for a retro/review process on the
service I think this is the best home, if we take it further we can look
at migrating that implementation here.