librato: fix authentication with retry and debug enabled
We've observed in our systems that retries never succeed, but instead
always return a `401 Unauthorized` after the failure.

It turns out when using both WithRetries and WithRequestDebugging the
authentication was being dropped before retry attempts were made.

This updates the reporter to preserve headers when capturing debug
information so that subsequent retries can succeed.