adds: cleanurl, nacl, and sharder
== Context

We've used these utilities pretty heavily within heroku and it's time to
extract it into the set of opensource tools that we'd like to use.

=== nacl

Standard authentication tools for encryption and signing.

=== cleanurl

A tool to extract the username/password from a URL. It was pretty useful
pre go 1.12 when creds weren't being scrubbed from errors.

=== sharder

A simple library for being able to choose the right shard based on a
given key.