Add mustcert (#70)
* Move mustcert from internal/ to pkg/

I find this to be useful for testing TLS related stuff elsewhere, so make this
code consumable in our other monorepi.

* mustcert: Slightly more flexible cert generation

Rename "hostname" to "commonName", to make what it is a little clearer for
non-serving usages. Add a optional function arg to the newCert and Leaf
methods, that takes a copy of the Cert that it can modify. This can be used to
add additional fields while keeping a simple method signature.

* mustcert: ec to p-256

Docker code rejects p-224, move to p-256. Slightly slower, but not enough to
matter

* update deps