testenv: improvements on defaults
== Context

Currently it was pointed to be a bit paranoid and try and avoid any
local installs.

While that's great for certain setups, that's not a good default for a
typical simple setup as well.

These new defaults allow a docker setup still, but also works well with
a typical brew install of redis / postgres.