Add Open Census agent exporting to the Service.
Don't start the open census agent exporter unless the Service is created
with EnableOpenCensusTracing. This allows for a level of opt-in.

All Open Census env variables are prefixed with `OC_`. Only a few
options are currently supported, but others could be added later.

Other cleanups were done for my own clarity while reading through the
code.