Re-work logging middleware to be compatible with Chi
* Previously, I added logging middleware that coudld be used by the Chi
router but did not interact with Chi.
* When we were testing our rollbar setup in Lyra Identity, we found that
the Panic recovery was not working within Chi router endpoints.
* This was because the Recovery middleware depends on the Logging
middleware that Chi provides, which we were not using because we were
using this handy Logging middleware from X instead.
* At this point, it seemed like a better solution than writing our own
recovery middleware for Chi was to implement the LogFormatter and
LogEntry interfaces so we can use the Chi middleware for logging and
panic recovery.
* This is the approach shown here https://github.com/go-chi/chi/blob/cca4135d8dddff765463feaf1118047a9e506b4a/_examples/logging/main.go
* Also used by Lyra Labs here https://github.com/heroku/lyra-labs/pull/212/
* Logging has been adapted to use the fields that other Heroku services
use for easier Spluking etc.