testing/testlog: add New function
The testlog package's function to get a test logger was `NewNullLogger`,
which was used to match the function name in logrus' hook test package.
Most folks using this package now, however, have not used the logrus
package, and there are no other kinds of test loggers in the package, so
using `New` instead reads better.

This also documents the package and adds examples.