cmdutil/service: support `service.New(nil)`
envdecode requires that at least one field in a provided config struct
match from the environment. This sometimes makes the standard service
awkward to work with, particularly when gradually migrating an existing
system.

It's common to see this while getting started:

```
var cfg struct {
  AppName string `env:"APP_NAME"` // mirror APP_NAME from svclog for envdecode
}
s := service.New(&cfg)
```

This updates `service.New` to accept nil for for the appConfig argument
to ease conversion of services.