workaround to pass explicit histogram options
The commit works around the disconnect between the OTEL metrics Meter
interface for creating histograms and the actual Aggregator used to
collect and report on the data. Presently, there is no way to pass
bucket definitions to the aggregator except globally for all histograms.

This commit resolves that by allowing for the use of a custome
AggregatorSelector that caches any options and looks them up when a new
Aggregator is requested by the OTEL controller.