Use RWMutex in OTEL provider (#214)
Currently, the `otel.Provider` uses a `sync.Mutex` to protect against race conditions on the maps that store instances of the different metric types. The mutex blocks to ensure the existence of a metric on metric modification. 

We can avoid fully blocking on reads for already existing metrics with a `sync.RWMutex`.